+++ src/components/icons/CrosshairIcon.vue
@@ -1,0 +1,33 @@
+<template>
+  <svg
+    id="CrosshairIcon"
+    xmlns="http://www.w3.org/2000/svg"
+    width="16"
+    height="16"
+    viewBox="0 0 16 16"
+  >
+    <g id="Gruppe_877" transform="translate(-2408 -282)">
+      <circle
+        id="Ellipse_100"
+        cx="3"
+        cy="3"
+        r="3"
+        transform="translate(2412.979 286.953)"
+        fill="currentColor"
+      />
+      <path
+        fill="currentColor"
+        id="Pfad_445"
+        d="M15.316,7.359l-1.36-.008A6,6,0,0,0,8.666,2.062V.73a.666.666,0,0,0-1.332,0V2.062a6,6,0,0,0-5.29,5.289L.683,7.359a.666.666,0,0,0,0,1.331s0,0,0-.015l1.355-.007a6,6,0,0,0,5.289,5.3v1.361a.666.666,0,0,0,1.332,0V13.965a6,6,0,0,0,5.289-5.3l1.354.007c0,.015,0,.015,0,.015a.666.666,0,0,0,0-1.331ZM8,12.673a4.66,4.66,0,1,1,4.663-4.66A4.666,4.666,0,0,1,8,12.673Z"
+        transform="translate(2407.979 281.936)"
+      />
+    </g>
+    <rect
+      id="size"
+      width="16"
+      height="16"
+      transform="translate(0 0)"
+      fill="none"
+    />
+  </svg>
+</template>