+++ src/uiConfig.js
@@ -29,6 +29,7 @@ import { ref } from 'vue';
  * @property {TextPageType} [dataProtection] - an option dataProtection, will show a link in the footer. Default title is 'footer.dataProtection.title'.
  * @property {string} [favicon] - the favicon to set
  * @property {string} [headerTitle] - the title to display in the tab of the browser
+ * @property {boolean} [showLocator] - an optional flag whether to show the Locator in the map.
  */
 
 /**